requirements.txt
rubik-cube
numpy
torch
gym
